+++ mowl/nn/__init__.py
@@ -3,3 +3,5 @@ from .el.elem.module import ELEmModule
 from .el.elbe.module import ELBEModule
 from .el.boxel.module import BoxELModule
 from .el.boxsquaredel.module import BoxSquaredELModule
+from .alc.module import ALCModule
+from .alc.falcon.module import FALCONModule